Rename validator packages
